Align cloudconfig extensions rendering with aws-operator
