+++ service/controller/v7/cloudconfig/types.go
@@ -2,23 +2,18 @@ package cloudconfig
 
 import (
 	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
-	"github.com/giantswarm/certs"
-
-	"github.com/giantswarm/azure-operator/client"
-	"github.com/giantswarm/azure-operator/service/controller/setting"
-	"github.com/giantswarm/azure-operator/service/controller/v7/key"
 )
+
+type templateData struct {
+	calicoAzureFileParams
+	cloudProviderConfFileParams
+	certificateDecrypterUnitParams
+	ingressLBFileParams
+}
 
 type calicoAzureFileParams struct {
 	Cluster    providerv1alpha1.Cluster
 	CalicoCIDR string
-}
-
-func newCalicoAzureFileParams(obj providerv1alpha1.AzureConfig, calicoCIDR string) calicoAzureFileParams {
-	return calicoAzureFileParams{
-		Cluster:    obj.Spec.Cluster,
-		CalicoCIDR: calicoCIDR,
-	}
 }
 
 type cloudProviderConfFileVMType string
@@ -39,46 +34,10 @@ type cloudProviderConfFileParams struct {
 	UseManagedIdentityExtension bool
 }
 
-func newCloudProviderConfFileParams(azure setting.Azure, azureConfig client.AzureClientSetConfig, obj providerv1alpha1.AzureConfig) cloudProviderConfFileParams {
-	return cloudProviderConfFileParams{
-		AADClientID:                 azureConfig.ClientID,
-		AADClientSecret:             azureConfig.ClientSecret,
-		EnvironmentName:             azure.EnvironmentName,
-		Location:                    azure.Location,
-		PrimaryScaleSetName:         key.WorkerVMSSName(obj),
-		ResourceGroup:               key.ResourceGroupName(obj),
-		RouteTableName:              key.RouteTableName(obj),
-		SecurityGroupName:           key.WorkerSecurityGroupName(obj),
-		SubnetName:                  key.WorkerSubnetName(obj),
-		SubscriptionID:              azureConfig.SubscriptionID,
-		TenantID:                    azureConfig.TenantID,
-		VnetName:                    key.VnetName(obj),
-		UseManagedIdentityExtension: azure.MSI.Enabled,
-	}
-}
-
 type certificateDecrypterUnitParams struct {
 	CertsPaths []string
 }
 
 type ingressLBFileParams struct {
 	ClusterDNSDomain string
-}
-
-func newCertificateDecrypterUnitParams(certFiles certs.Files) certificateDecrypterUnitParams {
-	var certsPaths []string
-
-	for _, file := range certFiles {
-		certsPaths = append(certsPaths, file.AbsolutePath)
-	}
-
-	return certificateDecrypterUnitParams{
-		CertsPaths: certsPaths,
-	}
-}
-
-func newIngressLBFileParams(obj providerv1alpha1.AzureConfig) ingressLBFileParams {
-	return ingressLBFileParams{
-		ClusterDNSDomain: key.ClusterDNSDomain(obj),
-	}
 }
