+++ service/controller/v7/templates/ignition/calico_azure.go
@@ -1,10 +1,6 @@
-package cloudconfig
+package ignition
 
-const (
-	calicoAzureFileName       = "/srv/calico-azure.yaml"
-	calicoAzureFileOwner      = "root:root"
-	calicoAzureFilePermission = 0600
-	calicoAzureFileTemplate   = `# Extra changes:
+const CalicoAzureResources = `# Extra changes:
 #  - Added resource limits to calico-node.
 #  - Added resource limits to install-cni.
 #  - Made install-cni initContainer.
@@ -599,202 +595,3 @@ subjects:
   namespace: kube-system
 ---
 `
-)
-
-const (
-	certFileOwner      = "root:root"
-	certFilePermission = 0600
-)
-
-const (
-	// Below we need separate path for master's kubelet as master machine
-	// is regular VM not running in VMSS. This may change soonish.
-
-	cloudProviderConfFileName       = "/etc/kubernetes/config/azure.yaml"
-	cloudProviderConfFileOwner      = "root:root"
-	cloudProviderConfFilePermission = 0600
-	cloudProviderConfFileTemplate   = `cloud: {{ .EnvironmentName }}
-tenantId: {{ .TenantID }}
-subscriptionId: {{ .SubscriptionID }}
-resourceGroup: {{ .ResourceGroup }}
-location: {{ .Location }}
-{{- if not .UseManagedIdentityExtension }}
-aadClientId: {{ .AADClientID }}
-aadClientSecret: {{ .AADClientSecret }}
-{{- end}}
-primaryScaleSetName: {{ .PrimaryScaleSetName }}
-subnetName: {{ .SubnetName }}
-securityGroupName: {{ .SecurityGroupName }}
-vnetName: {{ .VnetName }}
-vmType: vmss
-routeTableName: {{ .RouteTableName }}
-useManagedIdentityExtension: {{ .UseManagedIdentityExtension }}
-`
-)
-
-const (
-	defaultStorageClassFileName       = "/srv/default-storage-class.yaml"
-	defaultStorageClassFileOwner      = "root:root"
-	defaultStorageClassFilePermission = 0600
-	defaultStorageClassFileTemplate   = `apiVersion: storage.k8s.io/v1beta1
-kind: StorageClass
-metadata:
-  name: default
-  annotations:
-    storageclass.beta.kubernetes.io/is-default-class: "true"
-  labels:
-    kubernetes.io/cluster-service: "true"
-provisioner: kubernetes.io/azure-disk
-parameters:
-  kind: Managed
-  storageaccounttype: Premium_LRS
----
-apiVersion: storage.k8s.io/v1beta1
-kind: StorageClass
-metadata:
-  name: managed-premium
-  annotations:
-  labels:
-    kubernetes.io/cluster-service: "true"
-provisioner: kubernetes.io/azure-disk
-parameters:
-  kind: Managed
-  storageaccounttype: Premium_LRS
----
-apiVersion: storage.k8s.io/v1beta1
-kind: StorageClass
-metadata:
-  name: managed-standard
-  annotations:
-  labels:
-    kubernetes.io/cluster-service: "true"
-provisioner: kubernetes.io/azure-disk
-parameters:
-  kind: Managed
-  storageaccounttype: Standard_LRS
-`
-)
-
-const (
-	etcdMountUnitName     = "var-lib-etcd.mount"
-	etcdMountUnitTemplate = `[Unit]
-Description=Mounts disk to /var/lib/etcd
-Requires=format-etcd-disk.service
-After=format-etcd-disk.service
-Before=etcd3.service
-
-[Mount]
-What=/dev/disk/by-label/etcd
-Where=/var/lib/etcd
-Type=ext4
-
-[Install]
-WantedBy=multi-user.target
-`
-	etcdDiskFormatUnitName     = "format-etcd-disk.service"
-	etcdDiskFormatUnitTemplate = `[Unit]
-Description=Formats the disk drive
-Requires=dev-disk-azure-scsi1-lun{{ .LUNID }}.device
-After=dev-disk-azure-scsi1-lun{{ .LUNID }}.device
-
-[Service]
-Type=oneshot
-RemainAfterExit=yes
-Environment="LABEL=etcd"
-Environment="DEV=/dev/disk/azure/scsi1/lun{{ .LUNID }}"
-# Do not wipe the disk if it's already being used, so the etcd data is persistent across reboot.
-ExecStart=-/bin/bash -c "if ! findfs LABEL=$LABEL > /tmp/label.$LABEL; then wipefs -a -f $DEV && mkfs.ext4 -T news -F -L $LABEL $DEV && echo wiped; fi"
-
-[Install]
-WantedBy=multi-user.target
-`
-	dockerMountUnitName     = "var-lib-docker.mount"
-	dockerMountUnitTemplate = `[Unit]
-Description=Mounts disk to /var/lib/docker
-Requires=format-docker-disk.service
-After=format-docker-disk.service
-Before=docker.service
-
-[Mount]
-What=/dev/disk/by-label/docker
-Where=/var/lib/docker
-Type=xfs
-
-[Install]
-WantedBy=multi-user.target
-`
-	dockerDiskFormatUnitName     = "format-docker-disk.service"
-	dockerDiskFormatUnitTemplate = `[Unit]
-Description=Formats the disk drive
-Requires=dev-disk-azure-scsi1-lun{{ .LUNID }}.device
-After=dev-disk-azure-scsi1-lun{{ .LUNID }}.device
-
-[Service]
-Type=oneshot
-RemainAfterExit=yes
-Environment="LABEL=docker"
-Environment="DEV=/dev/disk/azure/scsi1/lun{{ .LUNID }}"
-ExecStart=-/bin/bash -c "wipefs -a -f $DEV && mkfs.xfs -L $LABEL $DEV"
-
-[Install]
-WantedBy=multi-user.target
-`
-)
-
-// TODO: This part should be removed when parameterized charts will be supported.
-const (
-	ingressLBFileName       = "/srv/k8s-ingress-loadbalancer.yaml"
-	ingressLBFileOwner      = "root:root"
-	ingressLBFilePermission = 0600
-	ingressLBFileTemplate   = `apiVersion: v1
-kind: Service
-metadata:
-  name: ingress-loadbalancer
-  namespace: kube-system
-  annotations:
-    external-dns.alpha.kubernetes.io/hostname: ingress.{{ .ClusterDNSDomain }}.
-  labels:
-    app: ingress-loadbalancer
-spec:
-  type: LoadBalancer
-  ports:
-  - name: http
-    port: 80
-  - name: https
-    port: 443
-  selector:
-    k8s-app: nginx-ingress-controller
-`
-	ingressLBUnitName     = "k8s-addons-ingress-loadbalancer.service"
-	ingressLBUnitTemplate = `[Unit]
-Description=Add Kubernetes load balancer manifest to addons
-ConditionPathExists=!/srv/k8s-ingress-loadbalancer.yaml.lock
-[Service]
-Type=oneshot
-ExecStart=/bin/sh -c "sed -i '/MANIFESTS=\"\"/a MANIFESTS=\"${MANIFESTS} k8s-ingress-loadbalancer.yaml\"' /opt/k8s-addons && touch /srv/k8s-ingress-loadbalancer.yaml.lock"
-[Install]
-WantedBy=multi-user.target
-`
-)
-
-// Certificate decryption
-const (
-	certDecrypterUnitName     = "certificate-decrypter.service"
-	certDecrypterUnitTemplate = `[Unit]
-Description=Certificate Decrypter
-Wants=k8s-setup-network-env.service
-After=k8s-setup-network-env.service
-Before=k8s-kubelet.service etcd3.service
-[Service]
-Type=oneshot
-EnvironmentFile=/etc/.enc/key
-EnvironmentFile=/etc/.enc/iv
-ExecStart=/bin/sh -c "\
-{{ range $index, $file := .CertsPaths -}}
-openssl enc -aes-256-cfb -d -K ${ENCRYPTION_KEY} -iv ${INITIAL_VECTOR} -in {{ $file }}.enc -out {{ $file }} ; \
-{{ end -}}
-"
- [Install]
-WantedBy=multi-user.target
-`
-)
