+++ service/controller/v7/templates/ignition/cloud_provider_conf.go
@@ -1,0 +1,19 @@
+package ignition
+
+const CloudProviderConf = `cloud: {{ .EnvironmentName }}
+tenantId: {{ .TenantID }}
+subscriptionId: {{ .SubscriptionID }}
+resourceGroup: {{ .ResourceGroup }}
+location: {{ .Location }}
+{{- if not .UseManagedIdentityExtension }}
+aadClientId: {{ .AADClientID }}
+aadClientSecret: {{ .AADClientSecret }}
+{{- end}}
+primaryScaleSetName: {{ .PrimaryScaleSetName }}
+subnetName: {{ .SubnetName }}
+securityGroupName: {{ .SecurityGroupName }}
+vnetName: {{ .VnetName }}
+vmType: vmss
+routeTableName: {{ .RouteTableName }}
+useManagedIdentityExtension: {{ .UseManagedIdentityExtension }}
+`
